+++ app/assets/stylesheets/application.scss
@@ -7,6 +7,7 @@
 @import "bootstrap/scss/bootstrap";
 @import "font-awesome-sprockets";
 @import "font-awesome";
+@import "animate.css/animate.css";
 
 // Your CSS partials
 @import "components/index";
