+++ app/assets/stylesheets/components/_modal.scss
@@ -25,3 +25,17 @@
     font-size: 36px;
   }
 }
+
+.modal-impact {
+  position: relative;
+  border-radius: 10px;
+  border: 5px solid $light-green;
+
+  #close {
+    position: absolute;
+    top: 8px;
+    right: 15px;
+    z-index: 2;
+    font-size: 36px;
+  }
+}
