--- app/views/action_completions/mass_create.js.erb
@@ -1,2 +1,0 @@
-// TODO: Animation once mass_create is done! confetti?
-console.log('fire');
