+++ app/views/action_settings/update.js.erb
@@ -1,0 +1,10 @@
+// AJAX for Dashboard Page, when clicking "+ HABIT"
+console.log('fire');
+action = document.getElementById('act-<%= @setting.action.id %>');
+console.log(action);
+
+action.classList.add('animated', 'bounceOutRight', 'delay-1s');
+
+action.addEventListener('animationend', function() { action.remove();})
+
+// addHabitBtn = action.querySelector('.add-habit-btn');
