+++ app/views/pages/_add_habit_link.html.erb
@@ -1,3 +1,3 @@
-<%= link_to action_setting_path(action.setting_for(current_user)), method: :patch do %>
+<%= link_to action_setting_path(action.setting_for(current_user)), method: :patch, remote: true do %>
   <p><i class="fas fa-plus"></i> HABIT</p>
 <% end %>
