+++ app/views/shared/_mass_create_button.html.erb
@@ -1,0 +1,3 @@
+<%= link_to action_completions_mass_create_path, method: :post, remote: true, id: 'habit-bottom-btn', class: display ? "" : "d-none" do %>
+  <i id='addon-checkable' class="fas fa-check-circle" ></i>
+<% end %>
