+++ package.json
@@ -3,6 +3,7 @@
   "private": true,
   "dependencies": {
     "@rails/webpacker": "4.2.2",
+    "animate.css": "^3.7.2",
     "bootstrap": "^4.4.1",
     "canvas-confetti": "^1.1.2",
     "chart.js": "^2.9.3",
